backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.0
pydantic==1.10.13
python-dotenv==1.0.0
